requirements.txt
fastapi==0.63.0
uvicorn==0.11.1
SQLAlchemy==1.4.12
databases==0.4.3
psycopg2-binary==2.8.6
PyJWT==2.1.0
passlib==1.7.4
bcrypt==3.2.0